requirements.txt
pandas
requests
psycopg2-binary
sqlalchemy
pandas_datareader
sidrapy
ipeadatapy
python-dotenv
pyspark
boto3
bs4
odfpy
openpyxl
basedosdados
python-calamine